+++ src/lib/buttons/Button.svelte
@@ -127,6 +127,9 @@
   on:change
   on:keydown
   on:keyup
+  on:touchstart
+  on:touchend
+  on:touchcancel
   on:mouseenter
   on:mouseleave>
   <slot />
@@ -165,6 +168,9 @@
   - on:change
   - on:keydown
   - on:keyup
+  - on:touchstart
+  - on:touchend
+  - on:touchcancel
   - on:mouseenter
   - on:mouseleave
   ## Example